Rename texture_zero.ts -> texture_zero.spec.ts
