+++ src/webgpu/api/operation/resource_init/check_texture/by_sampling.ts
@@ -8,7 +8,7 @@ import {
   getSingleDataType,
   getComponentReadbackTraits,
 } from '../../../../util/texture/texel_data.js';
-import { CheckContents } from '../texture_zero.js';
+import { CheckContents } from '../texture_zero.spec.js';
 
 export const checkContentsBySampling: CheckContents = (
   t,
